still need a little debugging
I have changed the name of the methods and I think the whole idea of having these start and stop tracking methods suggested is not correct. 
O also don't see the point of returning a boolean value after registering and deregistering if you are never going to use it so I am leaving the 'is tracking' to go in the magnifier because otherwise you will need to perform that kind of check twice, which is pointless.